Refactor IRandom and RandomWrapper
